Write to multiple registers is working properly
The write to multiple registers was fixed. The problem was with the
confusing system of adressing of f28335. When you write values to
address 0x8000 you will write an 32 bits value which will overwrite the
16 bits from 0x8000 and the other 16 bits are from 0x8001 (the high part
of the register). When you write something at 0x8001 it is like you are
writing at 0x8000. It means the changes are possible only at
*(address+2*i) adresses.